instrument was getting a lil big
