+++ src/index.js
@@ -7,7 +7,7 @@ const crypto = require('./modules/crypto');
 const ore = require('./tokens/ore');
 const oreStandardToken = require('./orestandardtoken');
 const rightsRegistry = require('./rightsregistry');
-
+const usageLog = require('./usagelog');
 
 class Orejs {
   constructor(config = {}) {
@@ -22,6 +22,7 @@ class Orejs {
     Object.assign(this, ore);
     Object.assign(this, oreStandardToken);
     Object.assign(this, rightsRegistry);
+    Object.assign(this, usageLog);
   }
 
   constructEos(config) {
