+++ src/instrument.js
@@ -1,8 +1,6 @@
 const APIM_CONTRACT_NAME = 'manager.apim';
 const INSTR_CONTRACT_NAME = 'instr.ore';
-const INSTR_USAGE_CONTRACT_NAME = 'usagelog.ore';
 const INSTR_TABLE_NAME = 'tokens';
-const LOG_COUNT_TABLE_NAME = 'counts';
 
 const ecc = require('eosjs-ecc');
 
@@ -56,20 +54,6 @@ async function getInstruments(oreAccountName, category = undefined, filters = []
   return rows;
 }
 
-async function getInstrumentsByRight(instrumentList, rightName) {
-  // Gets all the instruments with a particular right
-  const instruments = await instrumentList.filter(instrument => getRight(instrument, rightName) !== undefined);
-  return instruments;
-}
-
-async function getInstrumentByOwner(instrumentList, owner) {
-  // Get all the instruments with a particular owner
-  let instruments = [];
-  instruments = instrumentList.filter(instrument => instrument.owner === owner);
-
-  return instruments;
-}
-
 async function findInstruments(oreAccountName, activeOnly = true, category = undefined, rightName = undefined) {
   // Where args is search criteria could include (category, rights_name)
   // Note: this requires an index on the rights collection (to filter right names)
@@ -82,63 +66,6 @@ async function findInstruments(oreAccountName, activeOnly = true, category = und
   }
   const rows = await getInstruments.bind(this)(oreAccountName, category, filters);
   return rows;
-}
-
-async function getApiCallStats(instrumentId, rightName) {
-  // calls the usagelog contract to get the total number of calls against a particular right
-  const result = await this.eos.getTableRows({
-    code: INSTR_USAGE_CONTRACT_NAME,
-    json: true,
-    scope: instrumentId,
-    table: LOG_COUNT_TABLE_NAME,
-    limit: -1,
-  });
-
-  const rightProperties = {
-    totalApiCalls: 0,
-    totalCpuUsage: 0,
-  };
-
-  const rightObject = await result.rows.find(right => right.right_name === rightName);
-
-  if (rightObject !== undefined) {
-    rightProperties.totalApiCalls = rightObject.total_count;
-    rightProperties.totalCpuUsage = rightObject.total_cpu;
-  }
-
-  return rightProperties;
-}
-
-async function getRightStats(rightName, owner) {
-  // Returns the total cpu and api calls against a particular right across all the vouchers. If owner specified, then returns the toatal api calls and cpu usage for the owner.
-  let instruments;
-  let rightProperties;
-
-  const instrumentList = await this.getAllTableRows({
-    code: INSTR_CONTRACT_NAME,
-    scope: INSTR_CONTRACT_NAME,
-    table: INSTR_TABLE_NAME,
-    limit: -1,
-  });
-
-  instruments = await getInstrumentsByRight(instrumentList, rightName);
-
-  if (owner) {
-    instruments = await getInstrumentByOwner(instruments, owner);
-  }
-
-  // Get the total cpu calls and cpu count across all the instruments
-  const results = instruments.map(async (instrumentObject) => {
-    rightProperties = await getApiCallStats.bind(this)(instrumentObject.id, rightName);
-    return rightProperties;
-  });
-
-  const value = await Promise.all(results);
-
-  return {
-    totalCpuUsage: value.reduce((a, b) => a + parseFloat(b.totalCpuUsage), 0),
-    totalApiCalls: value.reduce((a, b) => a + parseFloat(b.totalApiCalls), 0),
-  };
 }
 
 async function createOfferInstrument(oreAccountName, offerInstrumentData, confirm = false) {
@@ -180,8 +107,6 @@ module.exports = {
   getRight,
   findInstruments,
   getInstruments,
-  getApiCallStats,
-  getRightStats,
   createOfferInstrument,
   createVoucherInstrument,
   signVoucher,
